openclaw-worker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.8.2
GitPython==3.1.43
